Link the period selection for the Dashboard, Card Feed and Monitoring screens (#6463)
