+++ ui/main/src/app/modules/share/timeline-buttons/timeline-buttons.component.html
@@ -22,7 +22,7 @@
         </li>
 
         <li *ngFor="let button of buttonList">
-          <a id="opfab-timeline-link-period-{{button.domainId}}" [ngClass]="button.selected ? 'link-selected' : 'link-unselect'" (click)="changeGraphConf(button)"
+          <a id="opfab-timeline-link-period-{{button.domainId}}" [ngClass]="button.selected ? 'link-selected' : 'link-unselect'" (click)="changeGraphConf(button, true)"
             translate>
             {{button.buttonTitle}}
           </a>
